requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
pytest==7.4.0
debugpy==1.8.0
requests==2.31.0